+++ src/contract/contract.ts
@@ -1,12 +1,14 @@
 import { Mixin } from 'ts-mixer'
 import { CsbContract } from './subcontracts/csb'
 import { LinkContract } from './subcontracts/link'
+import { NoteContract } from './subcontracts/note'
 import { ProfileContract } from './subcontracts/profile'
 import { RevisionContract } from './subcontracts/revision'
 
 const Contracts = Mixin(
   ProfileContract,
   LinkContract,
+  NoteContract,
   CsbContract,
   RevisionContract,
 )