+++ src/contract/subcontracts/profile.ts
@@ -1,7 +1,6 @@
 import { NIL_ADDRESS } from '../utils'
 import { BaseContract } from './base'
 import { autoSwitchMainnet } from '../decorators'
-import type { ProfileCreatedEvent } from '../abis/entry/types/Abi'
 import type { Result, Profile } from '../../types/contract'
 import { ProfileMetadata } from '../../types/metadata'
 import { Ipfs } from '../../ipfs'
@@ -11,9 +10,9 @@ export class ProfileContract extends BaseContract {
    * This creates a new profile for an address, and returns the ID of the newly created profile.
    * When the profile is the first profile created for an address, the address will be set as the primary profile.
    * @category Profile
-   * @param {string} owner - The Ethereum address of the profile owner.
-   * @param {string} handle - The handle of the profile you want to create.
-   * @param {string} metadataOrUri - The metadata or URI of the profile.
+   * @param owner - The Ethereum address of the profile owner.
+   * @param handle - The handle of the profile you want to create.
+   * @param metadataOrUri - The metadata or URI of the profile.
    * @returns The transaction hash and the profile ID.
    */
   @autoSwitchMainnet()
@@ -24,16 +23,9 @@ export class ProfileContract extends BaseContract {
   ): Promise<Result<string, true>> | never {
     this.validateHandleFormat(handle)
 
-    let uri
-    if (typeof metadataOrUri !== 'string') {
-      const metadata = metadataOrUri
-      if (!metadata.type) {
-        metadata.type = 'profile'
-      }
-      uri = await Ipfs.metadataToUri(metadata)
-    } else {
-      uri = metadataOrUri
-    }
+    const { uri } = await Ipfs.parseMetadataOrUri<ProfileMetadata>(
+      metadataOrUri,
+    )
 
     const tx = await this.contract.createProfile({
       to: owner,
@@ -45,10 +37,7 @@ export class ProfileContract extends BaseContract {
 
     const receipt = await tx.wait()
 
-    const parser = this.parseLog<ProfileCreatedEvent>(
-      receipt.logs,
-      'createProfile',
-    )
+    const parser = this.parseLog(receipt.logs, 'createProfile')
 
     return {
       data: parser.args.profileId.toString(),
@@ -59,8 +48,8 @@ export class ProfileContract extends BaseContract {
   /**
    * This changes a profile's handle.
    * @category Profile
-   * @param {string} profileId - The profile ID of the user you want to set the handle for.
-   * @param {string} handle - The handle you want to set.
+   * @param profileId - The profile ID of the user you want to set the handle for.
+   * @param handle - The handle you want to set.
    * @returns The transaction hash of the transaction that was sent to the blockchain.
    */
   @autoSwitchMainnet()
@@ -100,18 +89,10 @@ export class ProfileContract extends BaseContract {
         >
       >
     | never {
-    let uri
-    let metadata
-    if (typeof metadataOrUri !== 'string') {
-      metadata = metadataOrUri
-      if (!metadata.type) {
-        metadata.type = 'profile'
-      }
-      uri = await Ipfs.metadataToUri(metadata)
-    } else {
-      uri = metadataOrUri
-      metadata = await Ipfs.uriToMetadata<ProfileMetadata>(uri)
-    }
+    const { uri, metadata } = await Ipfs.parseMetadataOrUri<ProfileMetadata>(
+      metadataOrUri,
+      true,
+    )
 
     const tx = await this.contract.setProfileUri(profileId, uri)
     const receipt = await tx.wait()
@@ -129,7 +110,9 @@ export class ProfileContract extends BaseContract {
    * This is the same as {@link setProfileUri}
    * @category Profile
    */
-  setProfileMetadata = this.setProfileUri
+  async setProfileMetadata(profileId: string, metadata: ProfileMetadata) {
+    return this.setProfileUri(profileId, metadata)
+  }
 
   /**
    * This changes a profile's metadata (URI).
@@ -189,8 +172,8 @@ export class ProfileContract extends BaseContract {
   /**
    * This sets the social token for a profile
    * @category Profile
-   * @param {string} profileId - The profile ID of the user you want to set the social token for.
-   * @param {string} socialToken - The token address you want to set for the profile.
+   * @param profileId - The profile ID of the user you want to set the social token for.
+   * @param socialToken - The token address you want to set for the profile.
    * @returns The transaction hash of the transaction that was sent to the blockchain.
    */
   @autoSwitchMainnet()
@@ -209,7 +192,7 @@ export class ProfileContract extends BaseContract {
   /**
    * This sets the primary profile ID for the user.
    * @category Profile
-   * @param {string} profileId - The profile ID of the profile you want to set as primary.
+   * @param profileId - The profile ID of the profile you want to set as primary.
    * @returns The transaction hash of the transaction that was sent to the blockchain.
    */
   @autoSwitchMainnet()
@@ -227,7 +210,7 @@ export class ProfileContract extends BaseContract {
   /**
    * This returns the primary profile ID of the given address.
    * @category Profile
-   * @param {string} address - The address of the user you want to get the primary profile ID for.
+   * @param address - The address of the user you want to get the primary profile ID for.
    * @returns The profileId of the primary profile of the address.
    */
   @autoSwitchMainnet()
@@ -241,7 +224,7 @@ export class ProfileContract extends BaseContract {
   /**
    * This returns a boolean indicating whether the profileId is a primary profileId of an address.
    * @category Profile
-   * @param {string} profileId - The profile ID of the profile you want to check.
+   * @param profileId - The profile ID of the profile you want to check.
    * @returns A boolean value.
    */
   @autoSwitchMainnet()
@@ -257,18 +240,25 @@ export class ProfileContract extends BaseContract {
   /**
    * This returns a profile given its handle.
    * @category Profile
-   * @param {string} handle - The handle of the profile you want to get the content of.
+   * @param handle - The handle of the profile you want to get the content of.
    * @returns The profile with the given handle.
    */
   @autoSwitchMainnet()
   async getProfileByHandle(handle: string): Promise<Result<Profile>> | never {
     handle = handle.toLowerCase()
+
     const profile = await this.contract.getProfileByHandle(handle)
+
+    const metadata = profile.uri
+      ? await Ipfs.uriToMetadata<ProfileMetadata>(profile.uri)
+      : undefined
+
     return {
       data: {
         profileId: profile.profileId.toNumber().toString(),
         handle: profile.handle,
         uri: profile.uri,
+        metadata,
         socialToken: profile.socialToken,
         noteCount: profile.noteCount.toNumber(),
       },
@@ -278,17 +268,17 @@ export class ProfileContract extends BaseContract {
   /**
    * This returns a profile given its profileId.
    * @category Profile
-   * @param {string} profileId - The profile ID of the profile you want to get.
+   * @param profileId - The profile ID of the profile you want to get.
    * @returns The profile with the given profileId.
    */
   @autoSwitchMainnet()
   async getProfile(profileId: string): Promise<Result<Profile>> | never {
     const profile = await this.contract.getProfile(profileId)
 
-    let metadata: ProfileMetadata | undefined = undefined
-    if (profile.uri) {
-      metadata = await Ipfs.uriToMetadata(profile.uri)
-    }
+    const { metadata } = await Ipfs.parseMetadataOrUri<ProfileMetadata>(
+      profile.uri,
+      true,
+    )
 
     return {
       data: {
@@ -305,7 +295,7 @@ export class ProfileContract extends BaseContract {
   /**
    * This returns the handle of a profile.
    * @category Profile
-   * @param {string} profileId - The profileId of the profile you want to get the handle for.
+   * @param profileId - The profileId of the profile you want to get the handle for.
    * @returns The handle of the profile.
    */
   @autoSwitchMainnet()
@@ -319,7 +309,7 @@ export class ProfileContract extends BaseContract {
   /**
    * This returns the URI of a profile.
    * @category Profile
-   * @param {string} profileId - The profile ID of the profile you want to get the URI for.
+   * @param profileId - The profile ID of the profile you want to get the URI for.
    * @returns The URI of the profile.
    */
   @autoSwitchMainnet()
@@ -342,10 +332,7 @@ export class ProfileContract extends BaseContract {
   ): Promise<Result<Profile>> | never {
     const receipt = await this.contract.provider.getTransactionReceipt(txHash)
 
-    const parser = this.parseLog<ProfileCreatedEvent>(
-      receipt.logs,
-      'createProfile',
-    )
+    const parser = this.parseLog(receipt.logs, 'createProfile')
 
     const profileId = parser.args.profileId.toString()
     const result = await this.getProfile(profileId)
@@ -389,11 +376,13 @@ export class ProfileContract extends BaseContract {
 
   /**
    * This validates if a handle is in a correct format.
-   * @param {string} handle - The handle of the profile you want to get the social token for.
+   * @param handle - The handle of the profile you want to get the social token for.
    */
   private validateHandleFormat(handle: string): void | never {
-    if (handle.length >= 32) {
-      throw new Error(`Invalid handle: handle must be less than 32 characters.`)
+    if (handle.length >= 32 || handle.length <= 2) {
+      throw new Error(
+        `Invalid handle: handle must be between 3 and 31 characters.`,
+      )
     }
 
     if (!/^[a-z0-9\-\_]+$/.test(handle)) {